+++ default_www/frontend/modules/pages/widgets/sitemap.php
@@ -23,6 +23,9 @@ class FrontendPagesWidgetSitemap extends FrontendBaseWidget
 
 		// load template
 		$this->loadTemplate();
+
+		// assign sitemap navigation
+		$this->tpl->assign('widgetPagesNavigation', FrontendNavigation::getNavigationHTML('page', 0, null, array(), true));
 	}
 }